requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#

django-guardian==1.4.9
django-js-asset==1.0.0    # via django-mptt
django-mptt==0.9.0
django-tls-middleware==1.0.1
pytz==2017.3              # via django
six==1.11.0               # via django-guardian
